ProjectCards를 common으로 쓰는 UI 폴더로 이동 및 이름 변경
